cast doubles to float because CDash wants it
git-svn-id: https://svn.opendtect.org/od/trunk@31525 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiWellAttrib/uisynthtorealscale.cc
@@ -326,7 +326,7 @@ void uiSynthToRealScale::updSynthStats()
     uiHistogramDisplay& histfld = *synthstatsfld_->dispfld_;
     histfld.setData( vals.arr(), vals.size() );
     histfld.putN();
-    synthstatsfld_->updateSlider( histfld.getStatCalc().average() );
+    synthstatsfld_->updateSlider( (float)histfld.getStatCalc().average() );
 }
 
 
@@ -467,7 +467,7 @@ void uiSynthToRealScale::updRealStats()
     uiHistogramDisplay& histfld = *realstatsfld_->dispfld_;
     histfld.setData( coll.vals_.arr(), coll.vals_.size() );
     histfld.putN();
-    realstatsfld_->updateSlider( histfld.getStatCalc().average() );
+    realstatsfld_->updateSlider( (float)histfld.getStatCalc().average() );
     setScaleFld( 0 );
 }
